Bugfix: `app_dir` was incorrect!
