+++ app_builder/errors.py
@@ -1,0 +1,2 @@
+class ApplicationYamlError(Exception):
+    pass
