+++ app_builder/paths.py
@@ -1,7 +1,12 @@
+import fnmatch
 import os
+import re
 import sys
 import tempfile
 from pathlib import Path
+from typing import List
+
+from .errors import ApplicationYamlError
 
 installed_dir = Path(
     os.environ["LOCALAPPDATA"] if sys.platform == "win32" else "/opt",
@@ -27,3 +32,29 @@ def set_from_base(dirname: str | Path) -> Path:
 
 live_repo = set_from_base("live-repo")
 versions = set_from_base("versions")
+
+
+def iglob(p: str | Path, pattern: str) -> List[Path]:
+    rule = re.compile(fnmatch.translate(pattern), re.IGNORECASE)
+    return [f for f in Path(p).glob("*") if rule.match(f.name)]
+
+
+def get_app_base_directory(start_dir: Path) -> Path:
+    """
+    Travel up from the starting directory to find the application's base directory which contains 'application.yaml'.
+    """
+    d = start_dir.resolve()
+    err = ApplicationYamlError(
+        "Expected git repository with 'application.yaml' at base. "
+        "To initiate app-builder within the current repo, use `app-builder init`."
+    )
+    for i in range(1000):
+        if len(iglob(d, "application.yaml") + iglob(d, ".git")) == 2:
+            return d.resolve()
+
+        if d.parent == d:  # like "c:" == "c:"
+            raise err
+
+        d = d.parent
+
+    raise err
